[REL] 乱数処理の long化
乱数処理について int型で処理をしていたのを、long型に修正。
(ただし、Linux上で警告が出るかもしれない。いずれ確認出来れば良いが…)

+++ toscom/TEST/com_testtos.c
@@ -1480,8 +1480,8 @@ void test_random( void )
     startFunc( __func__ );
     int succeed = 0;
 
-    for( int i = 0;  i < 10;  i++ ) { com_printf( "%3d ", com_rand( 20 ) ); }
-    for( int i = 0;  i < 100; i++ ) {if( com_checkChance(50) ) { succeed++; }}
+    for( long i = 0;  i < 10;  i++ ) { com_printf( "%3ld ", com_rand( 20 ) ); }
+    for( long i = 0;  i < 100; i++ ) {if( com_checkChance(50) ) { succeed++; }}
     com_printf( "\n succeed = %d\n", succeed );
 }
